cashlab-api/requirements.txt
# === Core Framework ===
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
python-dotenv==1.2.1

# === Database ===
SQLAlchemy==2.0.41
aiosqlite==0.21.0
asyncpg==0.31.0
greenlet==3.2.4

# === Validation & Settings ===
pydantic==2.11.3
pydantic-settings==2.9.1

# === Authentication ===
python-jose==3.5.0
bcrypt==4.3.0
passlib==1.7.4
google-auth>=2.0.0
cryptography>=36.0.0

# === PDF Parsing ===
pdfplumber==0.11.6

# === HTTP Client ===
httpx==0.28.1

# === Testing ===
pytest==8.3.5
pytest-asyncio==0.26.0

# === Utilities ===
python-dateutil==2.9.0.post0
